requirements.txt
click>=8.3.1
dotenv>=0.9.9
InquirerPy>=0.3.4
loguru>=0.7.3
patool>=4.0.2
pre-commit>=4.5.0
protontricks>=1.13.1
psutil>=7.1.3
pydantic>=2.12.5
pyinstaller>=6.17.0
pyinstaller-hooks-contrib>=2025.10
pyyaml>=6.0.3
requests>=2.32.5
ruff==0.14.6
send2trash>=2.0.0
websockets>=15.0.1
